requirements.txt
altgraph==0.17
bsdiff4==1.1.5
certifi==2020.12.5
chardet==4.0.0
dsdev-utils==1.0.4
ed25519==1.5
future==0.18.2
idna==2.10
numpy==1.19.5
obd==0.7.1
pandas==1.2.1
pbr==5.5.0
pefile==2019.4.18
Pint==0.7.2
pyinstaller==4.0
pyinstaller-hooks-contrib==2020.8
pyserial==3.4
python-dateutil==2.8.1
pytz==2020.5
pywin32-ctypes==0.2.0
requests==2.25.1
six==1.15.0
stevedore==1.32.0
urllib3==1.26.2
# test speedtest-cli